requirements.txt
alembic==1.8.1
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Jsonpify==1.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
Flask-Mail
passlib==1.7.4
Pillow==9.2.0
PyJWT==2.5.0
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
flask-swagger-ui==4.11.1
celery==5.2.7
Flask-Mail==0.9.1 
redis==4.3.4
APscheduler==3.6.3